Add exceptions in thrift interfaces.
Change client and server code in accordance with new thrift exceptions.
Client now will return NotFound exception instead of null. Exceptions which
witness coding mistakes will be printed to log as error and must not be
ignored.

+++ db-services/db-diagram-service/src/main/java/com/qreal/wmp/db/diagram/server/DiagramDbServiceHandler.java
@@ -3,9 +3,7 @@ package com.qreal.wmp.db.diagram.server;
 import com.qreal.wmp.db.diagram.dao.DiagramDao;
 import com.qreal.wmp.db.diagram.model.Diagram;
 import com.qreal.wmp.db.diagram.model.Folder;
-import com.qreal.wmp.thrift.gen.DiagramDbService;
-import com.qreal.wmp.thrift.gen.TDiagram;
-import com.qreal.wmp.thrift.gen.TFolder;
+import com.qreal.wmp.thrift.gen.*;
 import org.springframework.context.ApplicationContext;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,40 +19,63 @@ public class DiagramDbServiceHandler implements DiagramDbService.Iface {
     }
 
     @Override
-    public long saveDiagram(TDiagram diagram) {
+    public long saveDiagram(TDiagram diagram) throws TIdAlreadyDefined {
+        if (diagram.isSetId()) {
+            throw new TIdAlreadyDefined("Diagram id not null. To save diagram you should not assign id to diagram.");
+        }
         return diagramDao.saveDiagram(new Diagram(diagram), diagram.getFolderId());
     }
 
     @Override
-    public TDiagram openDiagram(long diagramID) {
-        Diagram diagram = diagramDao.openDiagram(diagramID);
-        return diagram != null ? diagram.toTDiagram() : null;
+    public TDiagram openDiagram(long diagramId) throws TNotFound {
+        Diagram diagram = diagramDao.openDiagram(diagramId);
+        if (diagram == null) {
+            throw new TNotFound(String.valueOf(diagramId), "Diagram not found.");
+        }
+        return diagram.toTDiagram();
     }
 
     @Override
-    public void deleteDiagram(long diagramId) {
+    public void deleteDiagram(long diagramId) throws TNotFound {
+        if (!diagramDao.isExistsDiagram(diagramId)) {
+            throw new TNotFound(String.valueOf(diagramId), "Diagram to delete not found.");
+        }
         diagramDao.deleteDiagram(diagramId);
     }
 
     @Override
-    public void rewriteDiagram(TDiagram diagram) {
+    public void rewriteDiagram(TDiagram diagram) throws TNotFound, TIdNotDefined {
+        if (!diagram.isSetId()) {
+            throw new TIdNotDefined("Diagram id is null. To rewrite diagram you should specify id.");
+        }
+        if (!diagramDao.isExistsDiagram(diagram.getId())) {
+            throw new TNotFound(String.valueOf(diagram.getId()), "Diagram to rewrite not found.");
+        }
         diagramDao.rewriteDiagram(new Diagram(diagram));
     }
 
     @Override
-    public long createFolder(TFolder folder) {
+    public long createFolder(TFolder folder) throws TIdAlreadyDefined {
+        if (folder.isSetId()) {
+            throw new TIdAlreadyDefined("Folder id not null. To save folder you should not assign id to folder.");
+        }
         return diagramDao.createFolder(new Folder(folder));
     }
 
     @Override
-    public void deleteFolder(long folderId) {
+    public void deleteFolder(long folderId) throws TNotFound {
+        if (!diagramDao.isExistsFolder(folderId)) {
+            throw new TNotFound(String.valueOf(folderId), "Folder to delete not found.");
+        }
         diagramDao.deleteFolder(folderId);
     }
 
     @Override
-    public TFolder getFolderTree(String username) {
+    public TFolder getFolderTree(String username) throws TNotFound {
         Folder folder = diagramDao.getFolderTree(username);
-
-        return folder != null ? folder.toTFolder() : null;
+        if (folder == null) {
+            throw new TNotFound(username, "FolderTree for specified user not found.");
+        }
+        return folder.toTFolder();
     }
 }
